[ refac ] move the internal functions to a helper and add more test to move.
