+++ jest.config.json
@@ -1,3 +1,6 @@
 {
-  "testRegex": "((\\.|/*.)(test))\\.js?$"
+  "testRegex": "((\\.|/*.)(test))\\.js?$",
+  "globals": {
+    "firstMove": true
+  }
 }
